Remove group_vars/all.yml (#18)
* Use the actual user not root

I'm deploying this ansible script to users which are named "k8s". At the
moment, the ansible script assumes that you are root which is not always
the case. This copies the actual configuration file to the user it's
being deployed to rather than the root user.

* Remove group_vars/all.yml

It's common practice to try and put all variables in the hosts /
inventory file rather than in group_vars/ (so you only have to edit one
file such as inventory or hosts.sample)

For example, some very-large Ansible projects using this characteristic:

https://github.com/kubernetes-incubator/kubespray
https://github.com/openshift/openshift-ansible